Successful fix for regular peephole LSTMs
Now, given the same parameters and inputs, all of the regular LSTMs (with no normalizations) give the same outputs and gradients.

I created a notebook as a frame work for similar experiments. Next, I'm going to make a copy of this notebook and change the codes accordingly for normalized ones.