BN is working
I forgot to add `grad_cell` in the middle of calculations; that was the problem causing the anomaly in gradients when the sequence length was longer than 1.

There's still some errors in gradients, but considering their magnitudes, they are probably the results of precision errors in calculations. The error gets slightly bigger as the length of the sequence increases, but it should be negligible... hopefully.